requirements.txt
aiogram~=3.19.0
SQLAlchemy~=2.0.40
paramiko~=3.5.1
transliterate~=1.10.2
asyncpg~=0.30.0
python-dotenv~=1.1.0